changed routing from pages to app router
