+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 import Image from "next/image";
-import HeroButtons from "./components/herobuttons/HeroButton";
-import Browse from "./components/browse/Browse";
-import How from "./components/how/How";
-import Experience from "./components/experience/Experience";
+import HeroButtons from "@/components/herobuttons/HeroButton";
+import Browse from "@/components/browse/Browse";
+import How from "@/components/how/How";
+import Experience from "@/components/experience/Experience";
 
 export default function Home() {
   return (
