+++ src/components/how/How.tsx
